requirements.txt
python-dotenv
openai
langchain
langchain-openai
langgraph
deepagents
pydantic
httpx
ipykernel